pyproject.toml
[project]
name = "tehnesis-test-app"
version = "0.1.0"
description = "Tестовое задание для Технезис"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.19.0",
    "aiosqlite>=0.21.0",
    "pandas>=2.2.3",
    "patchright>=1.51.3",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
test = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: mark a test as a unit test",
    "integration: mark a test as an integration test",
    "slow: mark a test as slow (takes more than a few seconds)",
]
